Move FragmentStrictMode to common package to access package-private fields/methods
